refactor: SelectedCardInfo 타입을 miniGame.ts에 추가하고 관련 import 업데이트
